Refactor UpCast resolution out
